개인적 수정사항 적용
